backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
validators==0.22.0
SQLAlchemy==2.0.21
bcrypt==4.0.1
python-dateutil==2.8.2
email-validator==2.0.0.post2
limits==3.7.0

# ML and NLP dependencies
transformers==4.35.0
torch==2.1.0
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.2
newspaper3k==0.2.8
spacy==3.7.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0 